+++ bitprim/idiomatic/BlockList.cs
@@ -13,14 +13,14 @@ namespace Bitprim
             return BlockListNative.chain_block_list_construct_default();
         }
 
-        protected override Block GetNthNativeElement(uint n)
+        protected override Block GetNthNativeElement(UInt64 n)
         {
-            return new Block(BlockListNative.chain_block_list_nth(NativeInstance, (UInt64)n), false);
+            return new Block(BlockListNative.chain_block_list_nth(NativeInstance, n), false);
         }
 
-        protected override uint GetCount()
+        protected override UInt64 GetCount()
         {
-            return (uint) BlockListNative.chain_block_list_count(NativeInstance);
+            return BlockListNative.chain_block_list_count(NativeInstance);
         }
 
         protected override void AddElement(Block element)
